+++ test/rules/nestjs-example/element-types.spec.js
@@ -1,10 +1,10 @@
 const { ELEMENT_TYPES: RULE } = require("../../../src/constants/rules");
 const { createRuleTester, pathResolvers } = require("../../support/helpers");
-const { elementTypesErrorMessage } = require("../../support/messages");
+const { elementTypesErrorMessage, elementTypesNoRuleMessage } = require("../../support/messages");
 
 const rule = require(`../../../src/rules/${RULE}`);
 
-const test = (settings, options, { absoluteFilePath }, errorMessages) => {
+const test = (settings, options, { absoluteFilePath }, errorMessages, base) => {
   const ruleTester = createRuleTester(settings);
 
   ruleTester.run(RULE, rule, {
@@ -108,7 +108,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 0),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              0,
+              elementTypesNoRuleMessage({
+                file: "'app'",
+                dep: "'interface' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -120,7 +127,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 1),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              1,
+              elementTypesNoRuleMessage({
+                file: "'app'",
+                dep: "'controller' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -132,7 +146,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 2),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              2,
+              elementTypesNoRuleMessage({
+                file: "'app'",
+                dep: "'service' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -144,7 +165,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 3),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              3,
+              elementTypesNoRuleMessage({
+                file: "'module' with base '', feature 'core' and fileName 'core'",
+                dep: "'controller' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -156,7 +184,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 4),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              4,
+              elementTypesNoRuleMessage({
+                file: "'module' with base '', feature 'core' and fileName 'core'",
+                dep: "'service' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -168,7 +203,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 5),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              5,
+              elementTypesNoRuleMessage({
+                file: "'module' with base '', feature 'cats' and fileName 'cats'",
+                dep: "'interceptor' with base '', feature 'core' and fileName 'logging'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -180,7 +222,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 6),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              6,
+              elementTypesNoRuleMessage({
+                file: "'module' with base '', feature 'cats' and fileName 'cats'",
+                dep: "'interceptor' with base '', feature 'core' and fileName 'transform'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -192,7 +241,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 7),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              7,
+              elementTypesNoRuleMessage({
+                file: "'controller' with base '', feature 'core' and fileName 'core'",
+                dep: "'service' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -204,7 +260,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 8),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              8,
+              elementTypesNoRuleMessage({
+                file: "'controller' with base '', feature 'core' and fileName 'core'",
+                dep: "'dto' with base '', feature 'cats' and fileName 'create-cat'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -216,7 +279,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 9),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              9,
+              elementTypesNoRuleMessage({
+                file: "'controller' with base '', feature 'core' and fileName 'core'",
+                dep: "'interface' with base '', feature 'cats' and fileName 'cats'",
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -228,7 +298,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 10),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              10,
+              elementTypesNoRuleMessage({
+                file: "'controller' with base '', feature 'core' and fileName 'core'",
+                dep: `'model' with base '${base}', feature 'cats' and fileName 'persian-cat'`,
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -240,7 +317,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 11),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              11,
+              elementTypesNoRuleMessage({
+                file: "'controller' with base '', feature 'core' and fileName 'core'",
+                dep: `'model' with base '${base}', feature 'cats' and fileName 'siamese-cat'`,
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -252,7 +336,14 @@ const test = (settings, options, { absoluteFilePath }, errorMessages) => {
         options,
         errors: [
           {
-            message: elementTypesErrorMessage(errorMessages, 12),
+            message: elementTypesErrorMessage(
+              errorMessages,
+              12,
+              elementTypesNoRuleMessage({
+                file: "'controller' with base '', feature 'cats' and fileName 'cats'",
+                dep: `'model' with base '${base}', feature 'core' and fileName 'core'`,
+              })
+            ),
             type: "ImportDeclaration",
           },
         ],
@@ -361,7 +452,8 @@ test(
   },
   ruleOptions,
   pathResolvers("nestjs-example"),
-  {}
+  {},
+  "test/fixtures/nestjs-example"
 );
 
 test(
@@ -435,7 +527,13 @@ test(
   },
   ruleOptions,
   pathResolvers("nestjs-example"),
-  {}
+  {
+    12: elementTypesNoRuleMessage({
+      file: "'controller' with base '', feature 'cats' and fileName 'cats'",
+      dep: `'model' with base 'nestjs-example', feature 'core' and fileName 'core'`,
+    }),
+  },
+  ""
 );
 
 test(
@@ -509,5 +607,11 @@ test(
   },
   ruleOptions,
   pathResolvers("nestjs-example"),
-  {}
+  {
+    12: elementTypesNoRuleMessage({
+      file: "'controller' with base '', feature 'cats' and fileName 'cats'",
+      dep: `'model' with base '', feature 'core' and fileName 'core'`,
+    }),
+  },
+  "test/fixtures/nestjs-example"
 );