+++ test/support/messages.js
@@ -5,11 +5,16 @@ function errorMessage(errors, index, defaultMessage) {
   return errors[index] || defaultMessage;
 }
 
-function elementTypesErrorMessage(errors, index) {
-  return errorMessage(errors, index, ELEMENT_TYPES_DEFAULT_MESSAGE);
+function elementTypesErrorMessage(errors, index, defaultMessage) {
+  return errorMessage(errors, index, defaultMessage || ELEMENT_TYPES_DEFAULT_MESSAGE);
+}
+
+function elementTypesNoRuleMessage({ file, dep }) {
+  return `No rule allowing this dependency was found. File is of type ${file}. Dependency is of type ${dep}`;
 }
 
 module.exports = {
   ELEMENT_TYPES_DEFAULT_MESSAGE,
   elementTypesErrorMessage,
+  elementTypesNoRuleMessage,
 };